./tools -> aerospike_data
