+++ cmd/service/main.go
@@ -1,18 +1,14 @@
 package main
 
 import (
-	"errors"
 	"flag"
 	"fmt"
 	"os"
 	"os/exec"
-	"sort"
-	"time"
 
 	"github.com/2beens/serjtubincom/internal"
-	"github.com/2beens/serjtubincom/internal/aerospike"
 	"github.com/2beens/serjtubincom/internal/logging"
-	as "github.com/aerospike/aerospike-client-go"
+	"github.com/2beens/serjtubincom/tools"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -36,21 +32,21 @@ func main() {
 	flag.Parse()
 
 	if *aeroSetup {
-		if err := setupAeroDb(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
+		if err := tools.SetupAeroDb(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
 			fmt.Printf("aero setup failed: %s\n", err)
 			os.Exit(1)
 		}
 		fmt.Println("\naero setup completed")
 		os.Exit(0)
 	} else if *aeroDataFix {
-		if err := fixAerospikeData(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
+		if err := tools.FixAerospikeData(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
 			fmt.Printf("aero data fix failed: %s\n", err)
 			os.Exit(1)
 		}
 		fmt.Println("\naero data fix completed")
 		os.Exit(0)
 	} else if *aeroMessageIdCounterSet {
-		if err := fixAerospikeMessageIdCounter(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
+		if err := tools.FixAerospikeMessageIdCounter(*aeroNamespace, *aeroMessagesSet, *aeroHost, *aeroPort); err != nil {
 			fmt.Printf("aero message id counter fix / set failed: %s\n", err)
 			os.Exit(1)
 		}
@@ -119,216 +115,4 @@ func tryGetLastCommitHash() (string, error) {
 		return "", err
 	}
 	return string(stdout), nil
-}
-
-func setupAeroDb(namespace, set, host string, port int) error {
-	fmt.Println("staring aero setup ...")
-
-	aeroClient, err := as.NewClient(host, port)
-	if err != nil {
-		return fmt.Errorf("failed to create aero client: %w", err)
-	}
-
-	// TODO: maybe drop index first ?
-	// aeroClient.DropIndex(...)
-
-	if err := createBoardMessagesSecondaryIndex(aeroClient, namespace, set); err != nil {
-		return err
-	}
-
-	// other setup functions when/if needed:
-
-	return nil
-}
-
-func createBoardMessagesSecondaryIndex(aeroClient *as.Client, namespace, set string) error {
-	task, err := aeroClient.CreateIndex(
-		nil,
-		namespace,
-		set,
-		"id_index",
-		"id",
-		as.NUMERIC,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to get create index task: %w", err)
-	}
-
-	waitSecondsMax := 20
-	for i := 0; i < waitSecondsMax; i++ {
-		if done, err := task.IsDone(); err != nil {
-			fmt.Println(".")
-		} else if done {
-			break
-		}
-		time.Sleep(time.Second)
-	}
-
-	if err = <-task.OnComplete(); err != nil {
-		return fmt.Errorf("failed to create index: %w", err)
-	}
-
-	return nil
-}
-
-func fixAerospikeMessageIdCounter(namespace, set, host string, port int) error {
-	fmt.Printf("staring aero message id counter fix [%s : %s] ...\n", namespace, set)
-
-	aeroClient, err := as.NewClient(host, port)
-	if err != nil {
-		return fmt.Errorf("failed to create aero client: %w", err)
-	}
-
-	spolicy := as.NewScanPolicy()
-	spolicy.ConcurrentNodes = true
-	spolicy.Priority = as.LOW
-	spolicy.IncludeBinData = false
-
-	recs, err := aeroClient.ScanAll(spolicy, namespace, set)
-	if err != nil {
-		return err
-	}
-
-	count := 0
-	for range recs.Results() {
-		count++
-	}
-
-	fmt.Printf("trying to set message id counter to: %d\n", count+1)
-
-	metadataSet := set + "-metadata"
-	updatedIdCounter, err := setMessageIdCounter(namespace, metadataSet, aeroClient, count+1)
-	if err != nil {
-		return fmt.Errorf("failed to set message id counter: %w", err)
-	}
-
-	fmt.Printf("message id counter set to: %d\n", updatedIdCounter)
-
-	return nil
-}
-
-func setMessageIdCounter(namespace, set string, aeroClient *as.Client, increment int) (int, error) {
-	messageIdCounterKey := "message-id-counter"
-
-	key, err := as.NewKey(namespace, set, messageIdCounterKey)
-	if err != nil {
-		return -1, err
-	}
-
-	counterBin := as.NewBin("idCounter", increment)
-	record, err := aeroClient.Operate(nil, key, as.AddOp(counterBin), as.GetOp())
-	if err != nil {
-		return -1, fmt.Errorf("failed to call aero operate: %w", err)
-	}
-
-	counterRaw, ok := record.Bins["idCounter"]
-	if !ok {
-		log.Printf("\n%+v\n\n", record.Bins)
-		return -1, errors.New("id counter not existing")
-	}
-
-	counter, ok := counterRaw.(int)
-	if !ok {
-		return -1, errors.New("id counter not an integer")
-	}
-
-	return counter, nil
-}
-
-func fixAerospikeData(namespace, set, host string, port int) error {
-	fmt.Println("staring aero data fix ...")
-
-	aeroClient, err := as.NewClient(host, port)
-	if err != nil {
-		return fmt.Errorf("failed to create aero client: %w", err)
-	}
-
-	recordSet, err := aeroClient.ScanAll(nil, namespace, set)
-	if err != nil {
-		return fmt.Errorf("failed to scan all messages: %w", err)
-	}
-
-	var records []*as.Result
-	var messages []*internal.BoardMessage
-	for rec := range recordSet.Results() {
-		if rec.Err != nil {
-			return fmt.Errorf("get all messages, record error: %w", rec.Err)
-		}
-
-		m := internal.MessageFromBins(aerospike.AeroBinMap(rec.Record.Bins))
-		messages = append(messages, &m)
-		records = append(records, rec)
-	}
-
-	fmt.Printf("received %d messages from aerospike:\n", len(messages))
-	for i := range messages {
-		msg := messages[i]
-		fmt.Printf("%d: %s\n", msg.ID, time.Unix(msg.Timestamp, 0))
-	}
-
-	sort.Slice(messages, func(i, j int) bool {
-		return messages[i].Timestamp < messages[j].Timestamp
-	})
-
-	fmt.Println()
-	fmt.Println("------------------------------------------")
-	fmt.Println()
-
-	skipDelete := make(map[int64]bool)
-	for i := range messages {
-		message := messages[i]
-		fmt.Printf("saving message %d: %+v: %s - %s\n", i, time.Unix(message.Timestamp, 0), message.Author, message.Message)
-
-		bins := as.BinMap{
-			"id":        i,
-			"author":    message.Author,
-			"timestamp": message.Timestamp,
-			"message":   message.Message,
-		}
-
-		key, err := as.NewKey(namespace, set, i)
-		if err != nil {
-			return fmt.Errorf("failed to create a new message key: %w", err)
-		}
-
-		exists, err := aeroClient.Exists(nil, key)
-		if err != nil {
-			return fmt.Errorf("failed to check message existance of %d: %w", message.Timestamp, err)
-		}
-		if exists {
-			skipDelete[message.Timestamp] = true
-			continue
-		}
-
-		if err = aeroClient.Put(nil, key, bins); err != nil {
-			return fmt.Errorf("failed to save a message [%s] in aero: %w", key, err)
-		}
-	}
-
-	fmt.Println()
-	fmt.Println("------------------------------------------")
-	fmt.Println()
-
-	fmt.Println("deleting old records ...")
-	for i := range records {
-		r := records[i]
-		timestamp, ok := r.Record.Bins["timestamp"].(int)
-		if !ok {
-			fmt.Printf("failed to get timestamp of record %v\n", r)
-		}
-		if skipDelete[int64(timestamp)] {
-			fmt.Printf("skip deleting %d\n", timestamp)
-			continue
-		}
-
-		fmt.Printf("deleting: %s\n", r.Record.Key)
-		deleted, err := aeroClient.Delete(nil, r.Record.Key)
-		if err != nil {
-			fmt.Printf(" >>> failed to delete %s: %s\n", r.Record.Key, err)
-		} else {
-			fmt.Printf(" > found and deleted: %t\n", deleted)
-		}
-	}
-
-	return nil
 }
